# CUDA: mul_mat_vec_q for batch sizes > 1