On master the `mul_mat_vec_q` kernel only supports a batch size of 1. This PR implements support for batch sizes up to 8 in a minimally invasive way. In this range `mul_mat_vec_q` is universally faster. For larger batch sizes it depends on the quantization format. To keep things simple I am therefore only enabling the new implementation for batch sizes <= 8 (which should be enough for techniques like speculative decoding). I think the optimal solution would be to rewrite the `mul_mat_vec_q` kernel in a way that maximizes memory bandwidth (and also more optimization for the competing `mul_mat_q` kernels) but that is a larger project. As part of this PR I have also deduplicated the code for calling `mul_mat_vec_q`. On my systems the performance changes as follows:

| GPU        | Model             |     Batch size | Test     |     t/s master |     t/s PR |     Speedup |
| :--------- | :---------------- | -------------: | :------- | -------------: | ---------: | ----------: |
| RTX 3090   | llama 7B Q2_K_M   |              1 | pp512    |         103.75 |     102.80 |        0.99 |
| RTX 3090   | llama 7B Q2_K_M   |              2 | pp512    |          99.39 |     177.01 |        1.78 |
| RTX 3090   | llama 7B Q2_K_M   |              4 | pp512    |         195.77 |     248.26 |        1.27 |
| RTX 3090   | llama 7B Q2_K_M   |              8 | pp512    |         246.80 |     297.62 |        1.21 |
| RTX 3090   | llama 7B Q3_K_S   |              1 | pp512    |          98.36 |      97.37 |        0.99 |
| RTX 3090   | llama 7B Q3_K_S   |              2 | pp512    |          89.10 |     170.57 |        1.91 |
| RTX 3090   | llama 7B Q3_K_S   |              4 | pp512    |         175.77 |     243.07 |        1.38 |
| RTX 3090   | llama 7B Q3_K_S   |              8 | pp512    |         221.98 |     292.72 |        1.32 |
| RTX 3090   | llama 7B Q4_0     |              1 | pp512    |         131.80 |     130.71 |        0.99 |
| RTX 3090   | llama 7B Q4_0     |              2 | pp512    |         197.91 |     241.99 |        1.22 |
| RTX 3090   | llama 7B Q4_0     |              4 | pp512    |         386.23 |     371.20 |        0.96 |
| RTX 3090   | llama 7B Q4_0     |              8 | pp512    |         498.03 |     480.78 |        0.97 |
| RTX 3090   | llama 7B Q4_1     |              1 | pp512    |         124.75 |     123.22 |        0.99 |
| RTX 3090   | llama 7B Q4_1     |              2 | pp512    |         187.37 |     234.17 |        1.25 |
| RTX 3090   | llama 7B Q4_1     |              4 | pp512    |         363.74 |     379.89 |        1.04 |
| RTX 3090   | llama 7B Q4_1     |              8 | pp512    |         493.88 |     471.49 |        0.95 |
| RTX 3090   | llama 7B Q4_K_S   |              1 | pp512    |         124.42 |     123.73 |        0.99 |
| RTX 3090   | llama 7B Q4_K_S   |              2 | pp512    |         145.04 |     204.73 |        1.41 |
| RTX 3090   | llama 7B Q4_K_S   |              4 | pp512    |         283.69 |     285.35 |        1.01 |
| RTX 3090   | llama 7B Q4_K_S   |              8 | pp512    |         341.05 |     352.97 |        1.03 |
| RTX 3090   | llama 7B Q5_0     |              1 | pp512    |         114.12 |     112.97 |        0.99 |
| RTX 3090   | llama 7B Q5_0     |              2 | pp512    |          97.08 |     216.30 |        2.23 |
| RTX 3090   | llama 7B Q5_0     |              4 | pp512    |         190.99 |     354.03 |        1.85 |
| RTX 3090   | llama 7B Q5_0     |              8 | pp512    |         301.19 |     433.77 |        1.44 |
| RTX 3090   | llama 7B Q5_1     |              1 | pp512    |         109.46 |     108.28 |        0.99 |
| RTX 3090   | llama 7B Q5_1     |              2 | pp512    |         120.85 |     210.62 |        1.74 |
| RTX 3090   | llama 7B Q5_1     |              4 | pp512    |         236.95 |     327.31 |        1.38 |
| RTX 3090   | llama 7B Q5_1     |              8 | pp512    |         359.41 |     468.41 |        1.30 |
| RTX 3090   | llama 7B Q5_K_S   |              1 | pp512    |         112.22 |     111.32 |        0.99 |
| RTX 3090   | llama 7B Q5_K_S   |              2 | pp512    |         110.10 |     201.15 |        1.83 |
| RTX 3090   | llama 7B Q5_K_S   |              4 | pp512    |         216.51 |     283.01 |        1.31 |
| RTX 3090   | llama 7B Q5_K_S   |              8 | pp512    |         267.93 |     347.67 |        1.30 |
| RTX 3090   | llama 7B Q6_K     |              1 | pp512    |          90.84 |      90.19 |        0.99 |
| RTX 3090   | llama 7B Q6_K     |              2 | pp512    |         100.04 |     145.84 |        1.46 |
| RTX 3090   | llama 7B Q6_K     |              4 | pp512    |         197.02 |     236.99 |        1.20 |
| RTX 3090   | llama 7B Q6_K     |              8 | pp512    |         254.85 |     306.45 |        1.20 |
| RTX 3090   | llama 7B Q8_0     |              1 | pp512    |          87.15 |      86.56 |        0.99 |
| RTX 3090   | llama 7B Q8_0     |              2 | pp512    |         118.66 |     166.80 |        1.41 |
| RTX 3090   | llama 7B Q8_0     |              4 | pp512    |         232.79 |     292.20 |        1.26 |
| RTX 3090   | llama 7B Q8_0     |              8 | pp512    |         314.13 |     404.77 |        1.29 |
| RX 6800    | llama 7B Q2_K_M   |              1 | pp512    |          59.11 |      59.45 |        1.01 |
| RX 6800    | llama 7B Q2_K_M   |              2 | pp512    |          11.67 |      89.29 |        7.65 |
| RX 6800    | llama 7B Q2_K_M   |              4 | pp512    |          23.27 |     108.07 |        4.64 |
| RX 6800    | llama 7B Q2_K_M   |              8 | pp512    |          46.35 |     147.49 |        3.18 |
| RX 6800    | llama 7B Q3_K_S   |              1 | pp512    |          56.74 |      56.73 |        1.00 |
| RX 6800    | llama 7B Q3_K_S   |              2 | pp512    |          10.93 |      84.17 |        7.70 |
| RX 6800    | llama 7B Q3_K_S   |              4 | pp512    |          21.80 |     102.01 |        4.68 |
| RX 6800    | llama 7B Q3_K_S   |              8 | pp512    |          43.39 |     144.84 |        3.34 |
| RX 6800    | llama 7B Q4_0     |              1 | pp512    |          65.28 |      64.94 |        0.99 |
| RX 6800    | llama 7B Q4_0     |              2 | pp512    |          34.46 |     127.55 |        3.70 |
| RX 6800    | llama 7B Q4_0     |              4 | pp512    |          68.61 |     212.68 |        3.10 |
| RX 6800    | llama 7B Q4_0     |              8 | pp512    |         136.14 |     280.72 |        2.06 |
| RX 6800    | llama 7B Q4_1     |              1 | pp512    |          61.42 |      61.22 |        1.00 |
| RX 6800    | llama 7B Q4_1     |              2 | pp512    |          31.48 |     120.18 |        3.82 |
| RX 6800    | llama 7B Q4_1     |              4 | pp512    |          62.75 |     215.77 |        3.44 |
| RX 6800    | llama 7B Q4_1     |              8 | pp512    |         124.38 |     279.39 |        2.25 |
| RX 6800    | llama 7B Q4_K_S   |              1 | pp512    |          56.59 |      56.15 |        0.99 |
| RX 6800    | llama 7B Q4_K_S   |              2 | pp512    |          26.56 |      98.99 |        3.73 |
| RX 6800    | llama 7B Q4_K_S   |              4 | pp512    |          52.98 |     146.75 |        2.77 |
| RX 6800    | llama 7B Q4_K_S   |              8 | pp512    |         105.32 |     166.09 |        1.58 |
| RX 6800    | llama 7B Q5_0     |              1 | pp512    |          58.98 |      58.68 |        0.99 |
| RX 6800    | llama 7B Q5_0     |              2 | pp512    |          28.46 |     115.67 |        4.06 |
| RX 6800    | llama 7B Q5_0     |              4 | pp512    |          56.72 |     209.11 |        3.69 |
| RX 6800    | llama 7B Q5_0     |              8 | pp512    |         112.88 |     257.96 |        2.29 |
| RX 6800    | llama 7B Q5_1     |              1 | pp512    |          54.62 |      54.44 |        1.00 |
| RX 6800    | llama 7B Q5_1     |              2 | pp512    |          26.32 |     106.53 |        4.05 |
| RX 6800    | llama 7B Q5_1     |              4 | pp512    |          52.37 |     198.94 |        3.80 |
| RX 6800    | llama 7B Q5_1     |              8 | pp512    |         103.59 |     237.57 |        2.29 |
| RX 6800    | llama 7B Q5_K_S   |              1 | pp512    |          53.91 |      53.48 |        0.99 |
| RX 6800    | llama 7B Q5_K_S   |              2 | pp512    |          26.45 |      95.36 |        3.60 |
| RX 6800    | llama 7B Q5_K_S   |              4 | pp512    |          52.78 |     132.93 |        2.52 |
| RX 6800    | llama 7B Q5_K_S   |              8 | pp512    |         104.92 |     155.61 |        1.48 |
| RX 6800    | llama 7B Q6_K     |              1 | pp512    |          52.17 |      51.77 |        0.99 |
| RX 6800    | llama 7B Q6_K     |              2 | pp512    |          24.81 |     100.83 |        4.06 |
| RX 6800    | llama 7B Q6_K     |              4 | pp512    |          49.43 |     139.27 |        2.82 |
| RX 6800    | llama 7B Q6_K     |              8 | pp512    |          98.05 |     142.07 |        1.45 |
| RX 6800    | llama 7B Q8_0     |              1 | pp512    |          44.83 |      44.63 |        1.00 |
| RX 6800    | llama 7B Q8_0     |              2 | pp512    |          34.70 |      87.87 |        2.53 |
| RX 6800    | llama 7B Q8_0     |              4 | pp512    |          69.00 |     172.24 |        2.50 |
| RX 6800    | llama 7B Q8_0     |              8 | pp512    |         136.64 |     206.36 |        1.51 |
| P40        | llama 7B Q2_K_M   |              1 | pp512    |          30.88 |      30.91 |        1.00 |
| P40        | llama 7B Q2_K_M   |              2 | pp512    |          13.50 |      40.95 |        3.03 |
| P40        | llama 7B Q2_K_M   |              4 | pp512    |          26.84 |      64.87 |        2.42 |
| P40        | llama 7B Q2_K_M   |              8 | pp512    |          51.91 |      81.14 |        1.56 |
| P40        | llama 7B Q3_K_S   |              1 | pp512    |          28.45 |      28.43 |        1.00 |
| P40        | llama 7B Q3_K_S   |              2 | pp512    |          13.29 |      40.13 |        3.02 |
| P40        | llama 7B Q3_K_S   |              4 | pp512    |          26.39 |      64.51 |        2.44 |
| P40        | llama 7B Q3_K_S   |              8 | pp512    |          51.05 |      81.00 |        1.59 |
| P40        | llama 7B Q4_0     |              1 | pp512    |          54.71 |      54.66 |        1.00 |
| P40        | llama 7B Q4_0     |              2 | pp512    |          21.81 |      57.80 |        2.65 |
| P40        | llama 7B Q4_0     |              4 | pp512    |          43.41 |      86.93 |        2.00 |
| P40        | llama 7B Q4_0     |              8 | pp512    |          82.34 |     123.46 |        1.50 |
| P40        | llama 7B Q4_1     |              1 | pp512    |          54.15 |      54.03 |        1.00 |
| P40        | llama 7B Q4_1     |              2 | pp512    |          22.52 |      57.54 |        2.56 |
| P40        | llama 7B Q4_1     |              4 | pp512    |          44.53 |      91.01 |        2.04 |
| P40        | llama 7B Q4_1     |              8 | pp512    |          84.30 |     124.96 |        1.48 |
| P40        | llama 7B Q4_K_S   |              1 | pp512    |          50.84 |      50.94 |        1.00 |
| P40        | llama 7B Q4_K_S   |              2 | pp512    |          20.60 |      43.42 |        2.11 |
| P40        | llama 7B Q4_K_S   |              4 | pp512    |          40.71 |      72.69 |        1.79 |
| P40        | llama 7B Q4_K_S   |              8 | pp512    |          77.35 |      92.97 |        1.20 |
| P40        | llama 7B Q5_0     |              1 | pp512    |          47.09 |      46.98 |        1.00 |
| P40        | llama 7B Q5_0     |              2 | pp512    |          20.28 |      51.54 |        2.54 |
| P40        | llama 7B Q5_0     |              4 | pp512    |          40.47 |      82.53 |        2.04 |
| P40        | llama 7B Q5_0     |              8 | pp512    |          76.87 |     115.32 |        1.50 |
| P40        | llama 7B Q5_1     |              1 | pp512    |          47.61 |      47.50 |        1.00 |
| P40        | llama 7B Q5_1     |              2 | pp512    |          21.57 |      49.75 |        2.31 |
| P40        | llama 7B Q5_1     |              4 | pp512    |          42.78 |      84.25 |        1.97 |
| P40        | llama 7B Q5_1     |              8 | pp512    |          81.15 |     121.31 |        1.49 |
| P40        | llama 7B Q5_K_S   |              1 | pp512    |          38.87 |      38.90 |        1.00 |
| P40        | llama 7B Q5_K_S   |              2 | pp512    |          18.95 |      48.89 |        2.58 |
| P40        | llama 7B Q5_K_S   |              4 | pp512    |          37.52 |      67.09 |        1.79 |
| P40        | llama 7B Q5_K_S   |              8 | pp512    |          71.46 |      94.45 |        1.32 |
| P40        | llama 7B Q6_K     |              1 | pp512    |          35.00 |      35.03 |        1.00 |
| P40        | llama 7B Q6_K     |              2 | pp512    |          19.09 |      41.42 |        2.17 |
| P40        | llama 7B Q6_K     |              4 | pp512    |          37.11 |      58.58 |        1.58 |
| P40        | llama 7B Q6_K     |              8 | pp512    |          71.53 |      71.52 |        1.00 |
| P40        | llama 7B Q8_0     |              1 | pp512    |          33.72 |      33.71 |        1.00 |
| P40        | llama 7B Q8_0     |              2 | pp512    |          21.72 |      42.41 |        1.95 |
| P40        | llama 7B Q8_0     |              4 | pp512    |          42.93 |      56.08 |        1.31 |
| P40        | llama 7B Q8_0     |              8 | pp512    |          81.27 |     102.88 |        1.27 |

I did not test the new XS/XSS quants because I did not yet get around to setting them up.