@ggerganov 2024-02-06
This is looking very good! I'm doing some V100 tests and seeing similar gains

> To keep things simple I am therefore only enabling the new implementation for batch sizes <= 8 (which should be enough for techniques like speculative decoding).

Yes, I think 8 should be fine for most purposes

> I think the optimal solution would be to rewrite the mul_mat_vec_q kernel in a way that maximizes memory bandwidth 

I was thinking of writing a `ggml_cuda_op_dequantize_mul_mat_vec` kernel that dequantizes into shared memory and loads 16x16 fragments (padding with 0 what is beyond batch size). This will have constant speed across all batches up to 16, but it remains to be seen if it can outperform `mul_mat_vec_q` for very small batches. I'm thinking even if there is a tiny regression, it might be worth it, so that we have a single set of mat-vec kernels and consistent performance for bs <= 16