@JohannesGaessler 2024-02-06
>I was thinking of writing a ggml_cuda_op_dequantize_mul_mat_vec kernel that dequantizes into shared memory and loads 16x16 fragments (padding with 0 what is beyond batch size).

I've tried some improved implementations for both `dequantize_mul_mat_vec` and `mul_mat_vec_q` where the data is loaded into shared memory in a coalesced way and then only unpacked in shared memory but I was not able to make these faster than the implementation on master. I think one issue is that on a hardware level shared memory is just a portion of the regular cache that is manually managed. And since the data access patterns are very predictable VRAM (cache) -> SRAM -> registers just adds extra copies compared to VRAM (cache) -> registers. It may be better to have a kernel that loads the data into registers and then distributes the scales as needed using `__shfl_sync`. Or if you load the data asynchronously from VRAM to SRAM that may also work to increase bandwidth (but as I said before, `memcpy_async` causes annoying build issues.).

I think one issue with FP16 arithmetic will be shared memory limits. In principle, if you can fit the entire hidden state into shared memory then you should be able to write a kernel that needs to load it only once per streaming multiprocessor. Even for the smallest hidden state of 4096 values, if you have a batch size of 16 you would need 16\*4096\*2 = 131072 bytes to store it as FP16. But you would only need 73728 bytes to store it as q8_1 and then it would fit into the 102400 bytes of shared memory per SM on Ampere (the A100 has 167936 bytes per SM).

There are also issues with tail effects. Ideally you would distribute the rows as evenly as possible between SMs to maximize GPU utilization but tensor cores restrict you to multiples of 8/16/32 (depending on whether there are at least 32/16/8 hidden state columns). If you just use `__dp4a` for integer dot products you can assign rows with a granularity of 1.