@ggerganov 2024-02-06
Do you have some 13B models handy to see how are the results (I have just 7B on this V100 machine and it will take me some time to setup)?

I suspect that the results might not be so good for larger models.
For example, with Q8_0, the speed after BS=6 actually starts to degrade significantly (set the `S_TG` column):

```bash
LLAMA_CUBLAS=1 make -j batched-bench && ./batched-bench /mnt/llama.cpp/models/open-llama/7B-v2/ggml-model-q8_0.gguf 4800 0 99 0 50 100 1,2,3,4,5,6,7,8,16,32,64
```

```
|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|    50 |    100 |    1 |    150 |    0.079 |   635.44 |    1.311 |    76.26 |    1.390 |   107.91 |
|    50 |    100 |    2 |    300 |    0.076 |  1313.75 |    1.469 |   136.15 |    1.545 |   194.17 |
|    50 |    100 |    3 |    450 |    0.082 |  1823.60 |    1.688 |   177.76 |    1.770 |   254.24 |
|    50 |    100 |    4 |    600 |    0.095 |  2111.73 |    1.659 |   241.09 |    1.754 |   342.11 |
|    50 |    100 |    5 |    750 |    0.096 |  2592.66 |    2.010 |   248.78 |    2.106 |   356.08 |
|    50 |    100 |    6 |    900 |    0.116 |  2582.24 |    2.394 |   250.60 |    2.510 |   358.51 |
|    50 |    100 |    7 |   1050 |    0.127 |  2753.24 |    2.854 |   245.25 |    2.981 |   352.19 |
|    50 |    100 |    8 |   1200 |    0.139 |  2871.77 |    3.713 |   215.43 |    3.853 |   311.47 |
|    50 |    100 |   16 |   2400 |    0.282 |  2837.93 |    5.076 |   315.22 |    5.358 |   447.95 |
|    50 |    100 |   32 |   4800 |    0.582 |  2750.42 |    9.534 |   335.62 |   10.116 |   474.49 |
```

I'm thinking, the larger the model, the stronger this effect would be maybe