@slaren 2024-02-06
`llama-bench` shows that performance increases consistently with batch sizes 1-8, what is different with `batched-bench`? Is it possibly an outlier?