@ggerganov 2024-02-06
This is using `llama-bench`:

```bash
LLAMA_CUBLAS=1 make -j && ./llama-bench -ngl 99 -m models/openllama-7b-v2/ggml-model-q8_0.gguf -p 512 -b 1,2,3,4,5,6,7,8,16 -n 0
```

  Device 0: Tesla V100-PCIE-16GB, compute capability 7.0, VMM: yes
| model                          |       size |     params | backend    | ngl |    n_batch | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |          1 | pp 512     |     75.14 ± 0.11 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |          2 | pp 512     |    135.18 ± 0.45 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |          3 | pp 512     |    179.34 ± 0.44 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |          4 | pp 512     |    246.62 ± 0.70 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |          5 | pp 512     |    257.17 ± 0.21 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |          6 | pp 512     |    259.58 ± 0.39 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |          7 | pp 512     |    256.51 ± 0.20 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |          8 | pp 512     |    225.86 ± 0.13 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |         16 | pp 512     |    360.31 ± 0.76 |

build: dbb795b9 (2075)

Running now before this PR to see how it compares and will update here

Btw @slaren you might be running a wrong commit? 9392ebd is on `master`