@JohannesGaessler 2024-02-06
>I suspect that the results might not be so good for larger models.

>I'm thinking, the larger the model, the stronger this effect would be maybe

The size of the model should not make a difference. That only increases the number of blocks that the GPU works on but not the relative speed of the kernels. I suspect it's an issue related to GPU architecture where on Volta the number of registers per thread ends up too high; I'll look into adding launch bounds.