@slaren 2024-02-06
I also see a performance drop with batch sizes 6-8 with 13B Q8_0 compared to the previous master with a 3090 Ti.