@ggerganov 2024-02-06
> The size of the model should not make a difference.

The model size does seem to matter somehow because on 3090 it's faster for 7B Q8_0, but it is slower for 13B Q8_0 at BS 6-8