@JohannesGaessler 2024-02-06
The regression I'm seeing in mine and slaren's data is for batch size 4->5 and 6->7. 4->6, 6->8, and 4->8 is consistently faster. I'm currently profiling a few runs to make sure there are no other weird things going on at those batch sizes.