@ggerganov 2024-02-06
Yup, let's limit the `mul_mat_vec_q` kernel to bs <= 4 for now